+++ src/main/fr/formiko/graphisme/PanneauCarte.java
@@ -25,56 +25,28 @@ import java.awt.event.*;
 import javax.swing.*;
 
 public class PanneauCarte extends Panneau implements MouseListener{
-  private int tailleDUneCase; // entre 10 et 500.
-  private int tailleDUneCaseBase = 500;
   //private int espaceRéservéBas = 1; // permet de réservé un nombre de case défini d'espace.
   private int xCase; // nombre de case en X
   private int yCase; // nombre de case en Y
   private int posX; // position de la 1a case.
   private int posY;
   private int xTemp,yTemp;
-  private boolean imageIni;
-  private int scale = Image.SCALE_SMOOTH;
-  //image
-  private Image imgNull;
-  private Image tI1[];
-  private Image tI2[];
-  private Image tIF[][];
-  private Image tII[][];
-  private Image selectionnée; private Image fere;
-  private Image tG[][];
-  private Image tF[][];
-  private Graphics g;
-  private Image cNuageuse,cSombre;
-  private Image b[];
-  //ini
-  private Image imgNullIni;
-  private Image tI1Ini[];
-  private Image tI2Ini[];
-  private Image tIFIni[][];
-  private Image tIIIni[][];
-  private Image selectionnéeIni; private Image fereIni;
-  private Image tGIni[][];
-  private Image tFIni[][];
-  private Graphics gIni;
-  private Image cNuageuseIni,cSombreIni;
-  private Image bIni[];
 
   // CONSTRUCTEUR ---------------------------------------------------------------
   public PanneauCarte(){}
   public void construire(){
-    tailleDUneCase = Main.getTailleElementGraphique(100);
+    Main.getData().setTailleDUneCase(Main.getTailleElementGraphique(100));
     posX = 0; posY = 0;
     GCase gc = Main.getGc();
     xCase = gc.getNbrX();
     yCase = gc.getNbrY();
-    int xT = tailleDUneCase*xCase;
-    int yT = tailleDUneCase*yCase;
+    int xT = Main.getData().getTailleDUneCase()*xCase;
+    int yT = Main.getData().getTailleDUneCase()*yCase;
     this.setSize(xT,yT);
   }
   // GET SET --------------------------------------------------------------------
-  public int getTailleDUneCase(){return tailleDUneCase;}
-  public void setTailleDUneCase(int x){tailleDUneCase = x;}
+  public int getTailleDUneCase(){return Main.getData().getTailleDUneCase();}
+  public void setTailleDUneCase(int x){Main.getData().setTailleDUneCase(x);}
   public int getXCase(){ return xCase;}
   public void setXCase(int x){xCase = x;}
   public int getYCase(){ return yCase;}
@@ -103,8 +75,8 @@ public class PanneauCarte extends Panneau implements MouseListener{
       int x = Main.getPp().getWidth();
       int y = Main.getPp().getHeight();
       this.setSize(x,y);
-      xCase = math.min((x/tailleDUneCase)+1,gc.getNbrX()-posX);
-      yCase = math.min((y/tailleDUneCase)+1,gc.getNbrY()-posY);
+      xCase = math.min((x/Main.getData().getTailleDUneCase())+1,gc.getNbrX()-posX);
+      yCase = math.min((y/Main.getData().getTailleDUneCase())+1,gc.getNbrY()-posY);
       debug.débogage("Dimention du PanneauCarte en case : x="+xCase+" y="+yCase);
       debug.débogage("taille réèle du panneau de la carte : x="+this.getWidth()+", y="+this.getHeight());
       //dessin des cases :
@@ -117,54 +89,27 @@ public class PanneauCarte extends Panneau implements MouseListener{
       //dessin de la Fourmi selectionnée :
       if(Main.getFActuelle()!=null){
         Case c = Main.getFActuelle().getCCase().getContenu();
-        g.drawImage(selectionnée,(c.getX()-posX)*tailleDUneCase,(c.getY()-posY)*tailleDUneCase,this);
+        g.drawImage(Main.getData().getSelectionnee(),(c.getX()-posX)*Main.getData().getTailleDUneCase(),(c.getY()-posY)*Main.getData().getTailleDUneCase(),this);
       }
     }catch (Exception e) {
       erreur.erreur("Quelque chose d'imprévue est arrivé lors de l'affichage de PanneauCarte");
     }
     //Main.finCh("repaintDeLaCarte");
   }
-  public Image [] getScaledInstance(Image ti[],int dim, int b){
-    int lenr = ti.length;
-    Image r [] = new Image[lenr];
-    for (int i=0;i<lenr ;i++ ) {
-      if(b==0){//par défaut
-        r[i]=ti[i].getScaledInstance(dim, dim,scale);
-      }else if(b==1){//pour les fourmis.
-        int idEspece = 0;
-        int stade = i-3;
-        r[i]=ti[i].getScaledInstance(image.taille(idEspece, stade,dim), image.taille(idEspece, stade,dim),scale);
-      }else if(b==2){//pour les insectes
-        int idEspece = i+100;
-        int stade = 0;
-        r[i]=ti[i].getScaledInstance(image.taille(idEspece, stade,dim), image.taille(idEspece, stade,dim),scale);
-      }
-    }
-    return r;
-  }public Image [] getScaledInstance(Image ti[],int dim){return getScaledInstance(ti,dim,0);}
-  public Image [][] getScaledInstance(Image ti[][],int dim, int b){
-    int lenr = ti.length;
-    Image r [][] = new Image[lenr][];
-    for (int i=0;i<lenr ;i++ ) {
-      r[i]=getScaledInstance(ti[i],dim,b);
-    }
-    return r;
-  }public Image [][] getScaledInstance(Image ti[][],int dim){return getScaledInstance(ti,dim,0);}
-
   public void dessinerGrille(Graphics g){
     if(Main.getDessinerGrille()){
       for (int i=0;i<xCase+1 ;i++ ) {
-        int xT = tailleDUneCase*i;
-        g.drawLine(xT,0,xT,tailleDUneCase*yCase);
+        int xT = Main.getData().getTailleDUneCase()*i;
+        g.drawLine(xT,0,xT,Main.getData().getTailleDUneCase()*yCase);
       }
       for (int i=0;i<yCase+1 ;i++ ) {
-        int xT = tailleDUneCase*i;
-        g.drawLine(0,xT,tailleDUneCase*xCase,xT);
+        int xT = Main.getData().getTailleDUneCase()*i;
+        g.drawLine(0,xT,Main.getData().getTailleDUneCase()*xCase,xT);
       }
     }
   }
   public void repaintParciel(Case c){
-    peintImagePourCase(c,(Graphics2D)g);
+    peintImagePourCase(c,(Graphics2D) this.getGraphics());
   }
   public void peintImagePourCase(Case c,Graphics2D g){
     int x = c.getX(); int y = c.getY();
@@ -183,11 +128,11 @@ public class PanneauCarte extends Panneau implements MouseListener{
         }
         if (jo!=null){//si on a un joueur sélectionné.
           if ((x+posX)>=0 && (y+posY)>=0 && jo.getCaseNuageuse(x+posX,y+posY)){//si la case est invisible (nuageuse.)
-            g.drawImage(cNuageuse,xT,yT,this);
+            g.drawImage(Main.getData().getCNuageuse(),xT,yT,this);
             return true;//on ne dessine rien par dessus.
           }
         }else{//si pas de joueur selcetionné toute les cases sont nuageuse.
-          g.drawImage(cNuageuse,xT,yT,this);
+          g.drawImage(Main.getData().getCNuageuse(),xT,yT,this);
           return true;//on ne dessine rien par dessus.
         }
       }catch (Exception e) {
@@ -206,45 +151,35 @@ public class PanneauCarte extends Panneau implements MouseListener{
         fi=(Fourmi)jo.getFere().getGc().getDébut().getContenu();
       }catch (Exception e) {}
     }
-    int xT = x*tailleDUneCase; int yT = y*tailleDUneCase;
-    int xT2 = (x-posX)*tailleDUneCase; int yT2 = (y-posY)*tailleDUneCase;
+    int xT = x*Main.getData().getTailleDUneCase(); int yT = y*Main.getData().getTailleDUneCase();
+    int xT2 = (x-posX)*Main.getData().getTailleDUneCase(); int yT2 = (y-posY)*Main.getData().getTailleDUneCase();
     if(peintCaseNuageuse(x,y,g,xT,yT)){ return;}//si la case est nuageuse, on n'affichera rien d'autre dessus.
     byte ty = c.getType();
     CCreature ccrea = c.getGc().getDébut();
     CGraine ccg = c.getGg().getDébut();
-    int lenTIF = tIF[0].length+1;
+    int lenTIF = Main.getData().getTIF()[0].length+1;
     try {
       //le fond
       Image iTemp;
       if (ty==1) {
-        g.drawImage(tI2[0],xT,yT,this);
-        //iTemp = tI2[0];
+        g.drawImage(Main.getData().getTI2()[0],xT,yT,this);
       }else if (ty==2){
-        g.drawImage(tI2[1],xT,yT,this);
-        //iTemp = tI2[1];
+        g.drawImage(Main.getData().getTI2()[1],xT,yT,this);
       }else if (ty==3){
-        g.drawImage(tI1[1],xT,yT,this);
-        //iTemp = tI1[1];
+        g.drawImage(Main.getData().getTI1()[1],xT,yT,this);
       }else{
-        g.drawImage(imgNull,xT,yT,this);
-        //iTemp = imgNull;
+        g.drawImage(Main.getData().getImgNull(),xT,yT,this);
       }
-      /*try {
-        iTemp = iTemp.getScaledInstance(tailleDUneCase, tailleDUneCase,scale);
-        g.drawImage(iTemp,xT,yT,this);
-      }catch (Exception e) {
-        erreur.erreur("impossible de dessiner la case avec l'image redimentionnée.");
-      }*/
-      int tC10 = tailleDUneCase/10;int tC4 = tailleDUneCase/4;int tC2 = tailleDUneCase/2;
+      int tC10 = Main.getData().getTailleDUneCase()/10;int tC4 = Main.getData().getTailleDUneCase()/4;int tC2 = Main.getData().getTailleDUneCase()/2;
       // la fourmilière
       if (c.getFere()!=null){
-        g.drawImage(fere,xT+tC4,yT+tC4,this);
+        g.drawImage(Main.getData().getFere(),xT+tC4,yT+tC4,this);
         int tailleDuCercle = Main.getTailleElementGraphique(20);
-        drawRondOuRect(xT,yT,tailleDUneCase,g,c.getFere(),tailleDuCercle);
+        drawRondOuRect(xT,yT,Main.getData().getTailleDUneCase(),g,c.getFere(),tailleDuCercle);
         //affichage d'un rond de la couleur de la fere.
       }
       if(jo!=null && Main.getPartie().getCarte().getCasesSombres() && jo.getCaseSombre(x+posX,y+posY)){
-        g.drawImage(cSombre,xT,yT,this); // si les créatures sur la case ne sont pas visible.
+        g.drawImage(Main.getData().getCSombre(),xT,yT,this); // si les créatures sur la case ne sont pas visible.
       }else{
         // les graines
         int k=0;
@@ -253,7 +188,7 @@ public class PanneauCarte extends Panneau implements MouseListener{
             calculerXYTemp(xT,yT,k,c);k++;
             int dir = getDir((ObjetSurCarteAId)ccg.getContenu());
             try {
-              g.drawImage(tG[dir][ccg.getContenu().getType()],xTemp,yTemp,this);
+              g.drawImage(Main.getData().getTG()[dir][ccg.getContenu().getType()],xTemp,yTemp,this);
             }catch (Exception e) {}
             if(ccg.getContenu().getOuverte()){drawIcone(g,5,xT,yT,tC2);}
             else if(fi==null || ccg.getContenu().getDureté()<=fi.getDuretéMax()){drawIcone(g,4,xT,yT,tC2);}
@@ -271,19 +206,19 @@ public class PanneauCarte extends Panneau implements MouseListener{
             //System.out.println(ccrea.getContenu().getClass().equals(new Fourmi().getClass()));
             Fourmi f = ((Fourmi)ccrea.getContenu());
             if(f.getStade()==0){
-              g.drawImage(tIF[dir][math.min(f.getFourmiliere().getId()-1,lenTIF)],xTemp,yTemp,this);
+              g.drawImage(Main.getData().getTIF()[dir][math.min(f.getFourmiliere().getId()-1,lenTIF)],xTemp,yTemp,this);
             }else if(f.getStade()==-1){
-              g.drawImage(tF[dir][2],xTemp,yTemp,this);
+              g.drawImage(Main.getData().getTF()[dir][2],xTemp,yTemp,this);
             }else if(f.getStade()==-2){
-              g.drawImage(tF[dir][1],xTemp,yTemp,this);
+              g.drawImage(Main.getData().getTF()[dir][1],xTemp,yTemp,this);
             }else{ //stade == -3
-              g.drawImage(tF[dir][0],xTemp,yTemp,this);
+              g.drawImage(Main.getData().getTF()[dir][0],xTemp,yTemp,this);
             }
             insecte=false;
           }catch (Exception e) {
             try {
               Insecte i = (Insecte)(ccrea.getContenu());
-              g.drawImage(tII[dir][math.min(i.getType(),tII[dir].length)],xTemp,yTemp,this);
+              g.drawImage(Main.getData().getTII()[dir][math.min(i.getType(),Main.getData().getTII()[dir].length)],xTemp,yTemp,this);
             }catch (Exception e2) {erreur.erreur("impossible de dessiner l'image de la case : "+x+" "+y);}
           }
           //les icone
@@ -298,11 +233,15 @@ public class PanneauCarte extends Panneau implements MouseListener{
       erreur.erreur("impossible de dessiner l'image de la case : "+x+" "+y);
     }
   }
+  /**
+  *{@summary fonction that place ObjetSurCarteAId on the same Case.<br>}
+  *It modify PanneauCarte value xTemp and yTemp.
+  */
   public void calculerXYTemp(int xT, int yT, int k, Case c){
     int deplacementEnX=0;
     int deplacementEnY=0;
     //deplacement centré.
-    int espaceLibre = tailleDUneCase/5;
+    int espaceLibre = Main.getData().getTailleDUneCase()/5;
     if(Main.getPositionCase()==0 || (Main.getPositionCase()==2 && (k>3 || c.getGc().length()+c.getGg().length()==1))){//si on est en mode 2 et que la fourmi est seule ou qu'il y en a plus de 4.
       deplacementEnX=espaceLibre/2;
       deplacementEnY=espaceLibre/2;
@@ -343,7 +282,7 @@ public class PanneauCarte extends Panneau implements MouseListener{
   }
   public void drawIcone(Graphics g, int x, int xT, int yT, int tC2){
     if (!Main.getDessinerIcone()){ return;}
-    g.drawImage(b[x],xT+tC2,yT,this);
+    g.drawImage(Main.getData().getB()[x],xT+tC2,yT,this);
   }
   public int getDir(ObjetSurCarteAId obj){
     if (!Main.getElementSurCarteOrientéAprèsDéplacement()){return 0;}// si la direction de l'objet n'est pas prise en compte on cherche dans le tableau 0.
@@ -354,89 +293,7 @@ public class PanneauCarte extends Panneau implements MouseListener{
     return 3;
   }
 
-  //iniImage etc
-  public void chargerImages(){
-    debug.débogage("chargement des images a la bonne taille.");
-    if(!imageIni){
-      chargerImagesIni();
-      Main.débutCh();
-    }
-    int tailleFourmi = (tailleDUneCase*4)/5;
-    imgNull = imgNullIni.getScaledInstance(tailleDUneCase, tailleDUneCase,scale);
-    selectionnée = selectionnéeIni.getScaledInstance(tailleDUneCase, tailleDUneCase,scale);
-    tI1=getScaledInstance(tI1Ini, tailleDUneCase);
-    tI2=getScaledInstance(tI2Ini, tailleDUneCase);
-    tIF=getScaledInstance(tIFIni, tailleFourmi);
-    tII=getScaledInstance(tIIIni, tailleFourmi,2);//les insectes
-    tF=getScaledInstance(tFIni, tailleFourmi,1);//les Fourmis au différent stade.
-    tG=getScaledInstance(tGIni, tailleFourmi);
-    fere = fereIni.getScaledInstance(tailleDUneCase/2, tailleDUneCase/2,scale);
-    cNuageuse = cNuageuseIni.getScaledInstance(tailleDUneCase, tailleDUneCase,scale);
-    cSombre = cSombreIni.getScaledInstance(tailleDUneCase, tailleDUneCase,scale);
-    int lenb = bIni.length;
-    b=getScaledInstance(bIni,tailleDUneCase/2);
-    Main.finCh("chargerImages");
-  }
-  public void chargerImagesIni(){
-    if(!imageIni){
-      Main.débutCh();
-      imgNullIni = image.getImage("null").getScaledInstance(tailleDUneCaseBase, tailleDUneCaseBase,scale);
-      selectionnéeIni = image.getImage("selectionnee").getScaledInstance(tailleDUneCaseBase, tailleDUneCaseBase,scale);
-      chargerTI();
-      chargerTIF(Main.getNbrDeJoueur());
-      chargerTII();
-      chargerTF();
-      chargerTG();
-      fereIni = image.getImage("fourmiliere").getScaledInstance(tailleDUneCaseBase/2, tailleDUneCaseBase/2,scale);
-      cNuageuseIni = image.getImage("cNuageuse").getScaledInstance(tailleDUneCaseBase, tailleDUneCaseBase,scale);
-      cSombreIni = image.getImage("cSombre").getScaledInstance(tailleDUneCaseBase, tailleDUneCaseBase,scale);
-      bIni = image.getImages("b"); int lenb = bIni.length;
-      for (int i=0;i<lenb ;i++ ) {
-        bIni[i]=bIni[i].getScaledInstance(tailleDUneCaseBase/2, tailleDUneCaseBase/2,scale);
-      }
-      Main.finCh("chargerImagesIni");
-    }
-    imageIni=true;
-  }
-
-  public void chargerTI(){
-    tI1Ini = new Image [2]; tI1 = new Image [2];
-    tI1Ini[0]=image.getImage("terre").getScaledInstance(tailleDUneCaseBase, tailleDUneCaseBase,scale);
-    tI1Ini[1]=image.getImage("sable").getScaledInstance(tailleDUneCaseBase, tailleDUneCaseBase,scale);
-    tI2Ini = new Image [2]; tI2 = new Image [2];
-    tI2Ini[0]=image.getImage("herbe").getScaledInstance(tailleDUneCaseBase, tailleDUneCaseBase,scale);
-    tI2Ini[1]=image.getImage("mousse").getScaledInstance(tailleDUneCaseBase, tailleDUneCaseBase,scale);
-  }
-  public void chargerTIF(int nbrDeJoueur){
-    //if (!initialisationFX){initialisationFX=initialiserFX(nbrDeJoueur);}
-    tIFIni = chargerTX("F",nbrDeJoueur,(byte)0,1);
-  }public void chargerTIF(){ chargerTIF(12);}
-  public void chargerTII(){
-    tIIIni = chargerTX("I");
-  }
-  public void chargerTG(){
-    tGIni = chargerTX("graine");
-  }
-  public void chargerTF(){
-    tFIni = chargerTX("fourmi",3);
-  }
-  public Image [][] chargerTX(String nom, int x, byte y, int début){
-    Image tTemp [][] = image.getImagess(nom,x,(byte)début);
-    //mise a l'échelle.
-    int tailleFourmi = (tailleDUneCaseBase*4)/5;
-    int len1 = 4; int len2 = tTemp[0].length;
-    for (int j=0;j<len1 && tTemp[j]!=null; j++) {
-      for (int i=y;i<len2 ;i++ ) {
-        tTemp[j][i] = tTemp[j][i].getScaledInstance(tailleFourmi, tailleFourmi,scale);
-      }
-    }
-    return tTemp;
-  }
-  public Image [][] chargerTX(String nom, int x, byte y){ return chargerTX(nom,x,y,0);}
-  public Image [][] chargerTX(String sn, int x){ return chargerTX(sn,x,(byte)0);}
-  public Image [][] chargerTX(String sn){return chargerTX(sn,image.getNbrImages(sn));}
-
-
+  public void chargerImages(){Main.getData().chargerImages();}
 
 
   //sourie
